Modificaciones en el Front

+++ SCYGV/src/app/Components/Administrador/admin/admin.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import swal from'sweetalert2';
 
 @Component({
   selector: 'app-admin',
@@ -7,11 +8,21 @@ import { Router } from '@angular/router';
   styleUrls: ['./admin.component.css']
 })
 export class AdminComponent {
-  NombreUsuario : any = localStorage.getItem("Nombres")!;
+  NombreUsuario : any = localStorage.getItem("Nombres")! + " " + localStorage.getItem("Apellidos");
   constructor(private router:Router){}
   
-  CerrarSesion():void{
-    localStorage.clear();
-    this.router.navigateByUrl("/Home");
+  CerrarSesion():any{
+    swal.fire({
+      text: "¿Está seguro que desea eliminar?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: "Sí",
+      cancelButtonText: "Cancelar",
+  })
+  .then(resultado => {
+    if (resultado.value) {
+      localStorage.clear();
+      this.router.navigateByUrl("/Home");
+    }})
   }
 }
